backend/requirements.txt
Django>=4.0.0
djangorestframework>=3.10.0,<3.14.0
django-environ>=0.9.0
psycopg2==2.9.5
flake8>=5.0.0
Faker>=15.0.0
uwsgi>=2.0.18,<2.1.0
django-cors-headers